working connection to the DB, yaaaaaaaay

+++ src/main/java/com/littlemixrecipes/littlemix/services/CommentRepository.java
@@ -1,9 +1,9 @@
 package com.littlemixrecipes.littlemix.services;
 
-import org.springframework.data.repository.CrudRepository;
 
 import com.littlemixrecipes.littlemix.entities.CommentEntity;
+import org.springframework.data.repository.CrudRepository;
 
-public interface CommentRepository extends CrudRepository<CommentEntity, Long> {
+public interface CommentRepository extends CrudRepository<CommentEntity, Integer> {
 
 }
