+++ src/main/java/com/littlemixrecipes/littlemix/services/GradeRepository.java
@@ -1,9 +1,9 @@
 package com.littlemixrecipes.littlemix.services;
 
-import org.springframework.data.repository.CrudRepository;
 
 import com.littlemixrecipes.littlemix.entities.GradeEntity;
+import org.springframework.data.repository.CrudRepository;
 
-public interface GradeRepository extends CrudRepository<GradeEntity, Long> {
+public interface GradeRepository extends CrudRepository<GradeEntity, Integer> {
 	
 }
